added like and dislike, updated design and fixed some bugs

+++ MetioTube/profiles/signals.py
@@ -22,7 +22,11 @@ def user_created(sender, instance, created, **kwargs):
 
 @receiver(pre_save, sender=Profile)
 def change_profile_picture(sender, instance, **kwargs):
-    old_img = instance.__class__.objects.get(pk=instance.user_id).profile_picture
+    profile = instance.__class__.objects.filter(pk=instance.user_id).first()
+    if not profile:
+        return
+
+    old_img = profile.profile_picture
     if old_img:
         if old_img != instance.profile_picture:
             os.remove(old_img.path)
